refactor: 테스트용 MySQL, Redis 컨테이너 설정 클래스 분리
