+++ src/test/java/kr/hhplus/be/server/support/contanier/MySqlContainersConfig.java
@@ -1,15 +1,12 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.server.support.contanier;
 
 import jakarta.annotation.PreDestroy;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.utility.DockerImageName;
 
 @Configuration
-class TestcontainersConfiguration {
+class MySqlContainersConfig {
 
 	public static final MySQLContainer<?> MYSQL_CONTAINER;
